Refactor chain specs
