+++ node/src/chain_specs/mod.rs
@@ -1,0 +1,60 @@
+use cumulus_primitives_core::ParaId;
+use hex_literal::hex;
+use manta_primitives::{constants, currency::MA, AccountId, AuraId, Balance, Signature};
+use sc_chain_spec::{ChainSpecExtension, ChainSpecGroup};
+use sc_service::{ChainType, Properties};
+use serde::{Deserialize, Serialize};
+use sp_core::{crypto::UncheckedInto, sr25519, Pair, Public};
+use sp_runtime::traits::{IdentifyAccount, Verify};
+
+#[cfg(feature = "calamari")]
+pub mod calamari_pc;
+#[cfg(feature = "calamari")]
+pub use self::calamari_pc::*;
+#[cfg(feature = "manta-pc")]
+pub mod manta_pc;
+#[cfg(feature = "manta-pc")]
+pub use self::manta_pc::*;
+
+const ENDOWMENT: Balance = 1_000_000_000 * MA; // 10 endowment so that total supply is 10B
+const STAGING_TELEMETRY_URL: &str = "wss://api.telemetry.manta.systems/submit/";
+
+/// Helper function to generate a crypto pair from seed
+pub fn get_pair_from_seed<TPublic: Public>(seed: &str) -> <TPublic::Pair as Pair>::Public {
+	TPublic::Pair::from_string(&format!("//{}", seed), None)
+		.expect("static values are valid; qed")
+		.public()
+}
+
+/// Generate collator keys from seed.
+///
+/// This function's return type must always match the session keys of the chain in tuple format.
+pub fn get_collator_keys_from_seed(seed: &str) -> AuraId {
+	get_pair_from_seed::<AuraId>(seed)
+}
+
+/// The extensions for the [`ChainSpec`].
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, ChainSpecGroup, ChainSpecExtension)]
+#[serde(deny_unknown_fields)]
+pub struct Extensions {
+	/// The relay chain of the Parachain.
+	pub relay_chain: String,
+	/// The id of the Parachain.
+	pub para_id: u32,
+}
+
+impl Extensions {
+	/// Try to get the extension from the given `ChainSpec`.
+	pub fn try_get(chain_spec: &dyn sc_service::ChainSpec) -> Option<&Self> {
+		sc_chain_spec::get_extension(chain_spec.extensions())
+	}
+}
+
+type AccountPublic = <Signature as Verify>::Signer;
+/// Helper function to generate an account ID from seed
+pub fn get_account_id_from_seed<TPublic: Public>(seed: &str) -> AccountId
+where
+	AccountPublic: From<<TPublic::Pair as Pair>::Public>,
+{
+	AccountPublic::from(get_pair_from_seed::<TPublic>(seed)).into_account()
+}
