+++ node/src/lib.rs
@@ -1,2 +1,2 @@
-pub mod chain_spec;
+pub mod chain_specs;
 pub mod service;
