+++ node/src/manta-pc.rs
@@ -1,7 +1,7 @@
 //! Manta Parachain CLI
 #![warn(missing_docs)]
 
-mod chain_spec;
+mod chain_specs;
 #[macro_use]
 mod service;
 mod cli;
